+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Lights handle their own fading.
 - Light coordination separated into programs, e.g. Pulse, Reverse Pulse.
 
+### Changed
+- Project name: Light Array -> Lights. To accommodate matrix configuration.
+
 ## [0.4.0] - 2018-08-25
 ### Added
 - Light sources.